% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_make_server_fn.R
\name{filter_by_map_view_sf}
\alias{filter_by_map_view_sf}
\title{filter_by_map_view_sf}
\usage{
filter_by_map_view_sf(sf, bounds, col_chr)
}
\arguments{
\item{sf}{PARAM_DESCRIPTION}

\item{bounds}{PARAM_DESCRIPTION}

\item{col_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_crop}}
 \code{\link[dplyr]{pull}},\code{\link[dplyr]{filter}}
 \code{\link[rlang]{sym}}
}
