% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_server.R
\name{filter_if_var_exists}
\alias{filter_if_var_exists}
\title{filter_if_var_exists}
\usage{
filter_if_var_exists(tb, var_chr, var_val_xxx, cond_chr)
}
\arguments{
\item{tb}{PARAM_DESCRIPTION}

\item{var_chr}{PARAM_DESCRIPTION}

\item{var_val_xxx}{PARAM_DESCRIPTION}

\item{cond_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
