% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_launch_app.R
\name{launch_app}
\alias{launch_app}
\title{launch_app}
\usage{
launch_app(r_data_dir_chr, credentials_tb = NULL, shinyio_lgl = F)
}
\arguments{
\item{r_data_dir_chr}{PARAM_DESCRIPTION}

\item{credentials_tb}{PARAM_DESCRIPTION, Default: NULL}

\item{shinyio_lgl}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{shinyApp}}
}
