% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_ui.R
\name{make_app_ui}
\alias{make_app_ui}
\title{make_app_ui}
\usage{
make_app_ui()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{fluidPage}},\code{\link[shiny]{headerPanel}},\code{\link[shiny]{sidebarLayout}},\code{\link[shiny]{builder}},\code{\link[shiny]{htmlOutput}},\code{\link[shiny]{tabsetPanel}},\code{\link[shiny]{tabPanel}},\code{\link[shiny]{selectInput}},\code{\link[shiny]{actionButton}},\code{\link[shiny]{conditionalPanel}},\code{\link[shiny]{sliderInput}},\code{\link[shiny]{textInput}},\code{\link[shiny]{downloadButton}}
 \code{\link[shinythemes]{shinytheme}}
 \code{\link[shinyjs]{useShinyjs}}
 \code{\link[purrr]{flatten}}
}
