% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_make_ui_fn.R
\name{make_basic_ui_body_fn}
\alias{make_basic_ui_body_fn}
\title{make_basic_ui_fn}
\usage{
make_basic_ui_body_fn()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{fluidPage}},\code{\link[shiny]{headerPanel}},\code{\link[shiny]{htmlOutput}},\code{\link[shiny]{sliderInput}},\code{\link[shiny]{textInput}},\code{\link[shiny]{builder}},\code{\link[shiny]{radioButtons}},\code{\link[shiny]{downloadButton}},\code{\link[shiny]{selectInput}},\code{\link[shiny]{conditionalPanel}}
 \code{\link[shinythemes]{shinytheme}}
}
