% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_make_ui_fn.R
\name{make_basic_ui_fn}
\alias{make_basic_ui_fn}
\title{make_basic_ui_fn}
\usage{
make_basic_ui_fn(secure_lgl)
}
\arguments{
\item{secure_lgl}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shinymanager]{secure-app}}
}
