% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_ui.R
\name{make_inactivity_script_chr}
\alias{make_inactivity_script_chr}
\title{make_inactivity_script_chr}
\usage{
make_inactivity_script_chr()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
