% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_make_server_fn.R
\name{make_server_fn}
\alias{make_server_fn}
\title{make_server_fn}
\usage{
make_server_fn(r_data_dir_chr, credentials_tb = NULL)
}
\arguments{
\item{r_data_dir_chr}{PARAM_DESCRIPTION}

\item{credentials_tb}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shinymanager]{secure-app}},\code{\link[shinymanager]{check_credentials}}
 \code{\link[shiny]{reactiveValues}},\code{\link[shiny]{updateTabsetPanel}},\code{\link[shiny]{observeEvent}},\code{\link[shiny]{reactive}},\code{\link[shiny]{observe}},\code{\link[shiny]{renderUI}},\code{\link[shiny]{tag}},\code{\link[shiny]{selectInput}},\code{\link[shiny]{builder}},\code{\link[shiny]{checkboxGroupInput}},\code{\link[shiny]{actionButton}},\code{\link[shiny]{dateRangeInput}},\code{\link[shiny]{validate}},\code{\link[shiny]{sliderInput}},\code{\link[shiny]{HTML}},\code{\link[shiny]{downloadHandler}},\code{\link[shiny]{renderPrint}},\code{\link[shiny]{reactiveValuesToList}}
 \code{\link[shinyjs]{visibilityFuncs}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[springtides]{filter_if_var_exists}},\code{\link[springtides]{get_input_vec}},\code{\link[springtides]{get_disorder_choices_chr_vec}},\code{\link[springtides]{subset_vec_if_var_exists}},\code{\link[springtides]{get_age_range_choices_int_vec}},\code{\link[springtides]{get_input_ls}},\code{\link[springtides]{make_stat_choices_ls}},\code{\link[springtides]{make_area_name_chr}},\code{\link[springtides]{tf_stat_chr}}
 \code{\link[purrr]{reduce}},\code{\link[purrr]{map}},\code{\link[purrr]{flatten}}
 \code{\link[rlang]{sym}}
 \code{\link[stringr]{str_sub}},\code{\link[stringr]{str_replace}}
 \code{\link[lubridate]{period}},\code{\link[lubridate]{is.Date}},\code{\link[lubridate]{as_date}}
 \code{\link[rmarkdown]{render}},\code{\link[rmarkdown]{pdf_document}},\code{\link[rmarkdown]{html_document}},\code{\link[rmarkdown]{word_document}}
}
