% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_server.R
\name{secure_server}
\alias{secure_server}
\title{secure_server}
\usage{
secure_server(input, output, session)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{output}{PARAM_DESCRIPTION}

\item{session}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shinymanager]{secure-app}},\code{\link[shinymanager]{check_credentials}}
 \code{\link[shiny]{renderPrint}},\code{\link[shiny]{reactiveValuesToList}}
}
