% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_server.R
\name{server}
\alias{server}
\title{server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{output}{PARAM_DESCRIPTION}

\item{session}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[springtides]{c("get_input_vec", "get_input_vec")}},\code{\link[springtides]{c("get_disorder_choices_chr_vec", "get_disorder_choices_chr_vec")}},\code{\link[springtides]{c("get_age_range_choices_int_vec", "get_age_range_choices_int_vec")}},\code{\link[springtides]{c("make_area_name_chr", "make_area_name_chr")}},\code{\link[springtides]{c("tf_stat_chr", "tf_stat_chr")}}
 \code{\link[shiny]{reactiveValues}},\code{\link[shiny]{updateTabsetPanel}},\code{\link[shiny]{observeEvent}},\code{\link[shiny]{observe}},\code{\link[shiny]{reactive}},\code{\link[shiny]{renderUI}},\code{\link[shiny]{tag}},\code{\link[shiny]{selectInput}},\code{\link[shiny]{builder}},\code{\link[shiny]{checkboxGroupInput}},\code{\link[shiny]{actionButton}},\code{\link[shiny]{dateRangeInput}},\code{\link[shiny]{validate}},\code{\link[shiny]{sliderInput}},\code{\link[shiny]{HTML}},\code{\link[shiny]{downloadHandler}}
 \code{\link[shinyjs]{visibilityFuncs}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[purrr]{reduce}},\code{\link[purrr]{map}}
 \code{\link[rlang]{sym}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[stringr]{str_sub}},\code{\link[stringr]{str_replace}}
 \code{\link[lubridate]{is.Date}},\code{\link[lubridate]{as_date}}
 \code{\link[pkgload]{system.file}}
 \code{\link[rmarkdown]{render}},\code{\link[rmarkdown]{pdf_document}},\code{\link[rmarkdown]{word_document}}
 \code{\link[knitrBootstrap]{bootstrap_document}}
}
