% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_server.R
\name{source_fn_contents}
\alias{source_fn_contents}
\title{source_fn_contents}
\usage{
source_fn_contents(fn)
}
\arguments{
\item{fn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
