% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_server.R
\name{subset_vec_if_var_exists}
\alias{subset_vec_if_var_exists}
\title{subset_vec_if_var_exists}
\usage{
subset_vec_if_var_exists(chr_vec, var_val_chr, fn)
}
\arguments{
\item{chr_vec}{PARAM_DESCRIPTION}

\item{var_val_chr}{PARAM_DESCRIPTION}

\item{fn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{exec}}
}
