% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_launch.R
\name{launch_app}
\alias{launch_app}
\title{launch_app}
\usage{
launch_app(r_data_dir_chr)
}
\arguments{
\item{r_data_dir_chr}{A character string}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{getShinyOption}},\code{\link[shiny]{shinyApp}}
}
